add index sorters and extract 'read' logic
